updating tool names: append_node -> append_component merge -> merge_networks
