+++ src/konnektor/__init__.py
@@ -11,8 +11,9 @@ from .network_planners import (MaximalNetworkGenerator,
                                StarrySkyNetworkGenerator,
                                MstConcatenator,
                                )
-from .network_tools import concatenate, merge, append_node, \
-    delete_transformation, delete_component
+
+from .network_tools import merge_networks, concatenate_networks, \
+    append_component, delete_component, delete_transformation
 
 from . import network_analysis
 from .visualization import draw_ligand_network
