+++ src/konnektor/network_tools/__init__.py
@@ -1,6 +1,6 @@
 from .clustering.charge_clustering import ChargeClusterer
 from .clustering.component_diversity_clustering import \
     ComponentsDiversityClusterer
-from .concatenate import concatenate_networks, append_node
-from .merge import merge_two_networks
-from .misc import delete_transformation, delete_component
+
+from .network_handling import merge_networks, concatenate_networks, \
+    append_component, delete_component, delete_transformation
