+++ src/konnektor/network_tools/network_handling/__init__.py
@@ -1,0 +1,6 @@
+# This code is part of OpenFE and is licensed under the MIT license.
+# For details, see https://github.com/OpenFreeEnergy/konnektor
+
+from .concatenate import concatenate_networks, append_component
+from .delete import delete_transformation, delete_component
+from .merge import merge_networks, merge_two_networks
