+++ src/konnektor/network_tools/network_handling/misc.py
@@ -1,0 +1,24 @@
+# This code is part of OpenFE and is licensed under the MIT license.
+# For details, see https://github.com/OpenFreeEnergy/konnektor
+
+from gufe import Component
+from gufe import LigandNetwork
+
+
+def cyclize_around_component(network: LigandNetwork,
+                             node: Component) -> LigandNetwork:
+    """
+        TODO: Implement this
+        Not Implemented!
+
+    Parameters
+    ----------
+    network
+    node
+
+    Returns
+    -------
+
+    """
+    # TODO: Implement this
+    raise NotImplementedError()
